// src/main/model/game/ScarabSpinTypes.ts
export enum Symbol {
  RED_HEART = 'ace',
  YELLOW_UPSIDE_DOWN_TRIANGE = 'one',
  RED_RHOMBUS = 'two',
  PURPLE_HENTAGON = 'three',
  GREEN_OCTAGON = 'four',
  BLUE_CHIP = 'five',
  SCARAB_MUMMY = 'nine',
  GOLD_COIN = 'ten',
  GREEN_RHOMBUS = 'jack',
  BLUE_SPADE = 'queen',
  PURPLE_CLUB = 'king',
  SCATTER = 'scatter',
  WILD = 'wild'
}